new routes, new relations, edit controllers

+++ app/Http/Controllers/ProfileController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Profile;
+use App\Models\User;
 use Illuminate\Http\Request;
 
 class ProfileController extends Controller
@@ -14,9 +15,14 @@ class ProfileController extends Controller
      */
     public function index()
     {
+        $profiles = Profile::all();
+        foreach ($profiles as $key => $profile) {
+            $sup = User::find($profile->tier);
+            $profile->superviseur = $sup;
+        }
         return response()->json([
             'status' => 'true',
-            'data' => Profile::all()
+            'data' => $profiles
         ]);
     }
 
@@ -61,6 +67,8 @@ class ProfileController extends Controller
     public function show($id)
     {
         $profile = Profile::find($id);
+        $sup = User::find($profile->tier);
+        $profile->superviseur = $sup;
         if ($profile != null) {
             return response()->json([
                 'status' => 'true',
@@ -93,7 +101,7 @@ class ProfileController extends Controller
         $profile->tier = $request->tier;
         $profile->extern_clock = $request->extern_clock;
         $profile->save();
-        if ($profile != null) {
+        if ($profile->wasChanged()) {
             return response()->json([
                 'status' => 'true',
                 'data' => $profile
@@ -115,17 +123,24 @@ class ProfileController extends Controller
     public function destroy($id)
     {
         $profile = Profile::find($id);
-        $profile->delete();
-        if ($profile->deleted_at != null) {
-            return response()->json([
-                'status' => 'true',
-                'message' => 'deleted',
-            ]);
-        } else {
+        if ($profile->has('user')) {
             return response()->json([
                 'status' => 'false',
-                'message' => 'not deleted',
+                'message' => 'Ce profil ne peut être supprimé car est lié à des utilisateurs.',
             ]);
+        } else {
+            $profile->delete();
+            if ($profile->deleted_at != null) {
+                return response()->json([
+                    'status' => 'true',
+                    'message' => 'deleted',
+                ]);
+            } else {
+                return response()->json([
+                    'status' => 'false',
+                    'message' => 'not deleted',
+                ]);
+            }
         }
         
     }
